fix dupe activation bug
